+++ packages/react-dom/src/Link.tsx
@@ -25,6 +25,7 @@ export interface LinkProps
   onClick?: (e: React.MouseEvent<HTMLElement>) => void;
   anchor?: React.ReactType;
   children: NavigatingChildren | React.ReactNode;
+  forward?: React.AnchorHTMLAttributes<HTMLAnchorElement>;
 }
 
 interface BaseLinkProps extends LinkProps {
@@ -37,6 +38,7 @@ interface LinkState {
 }
 
 let hasWarnedTo = false;
+let hasWarnedForward = false;
 
 class BaseLink extends React.Component<BaseLinkProps, LinkState> {
   removed: boolean;
@@ -46,12 +48,17 @@ class BaseLink extends React.Component<BaseLinkProps, LinkState> {
   };
 
   shouldComponentUpdate(nextProps: BaseLinkProps, nextState: LinkState) {
-    const { params: nextParams, ...nextRest } = nextProps;
-    const { params: currentParams, ...currentRest } = this.props;
+    const { params: nextParams, forward: nextForward, ...nextRest } = nextProps;
+    const {
+      params: currentParams,
+      forward: currentForward,
+      ...currentRest
+    } = this.props;
     return (
       !shallowEqual(nextState, this.state) ||
       !shallowEqual(nextParams, currentParams) ||
-      !shallowEqual(nextRest, currentRest)
+      !shallowEqual(nextRest, currentRest) ||
+      !shallowEqual(nextForward, currentForward)
     );
   }
 
@@ -102,6 +109,7 @@ class BaseLink extends React.Component<BaseLinkProps, LinkState> {
       router,
       forwardedRef,
       children,
+      forward,
       ...rest
     } = this.props;
     if (process.env.NODE_ENV !== "production") {
@@ -112,8 +120,20 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
 
 <Link name="Route Name">...</Link>`);
       }
+
+      if (!hasWarnedForward && Object.keys(rest).length > 0) {
+        hasWarnedForward = true;
+        console.warn(`Deprecation warning:
+Passing additional props to a <Link> will no longer be forwarded to the rendered component in v2.
+
+Instead, please use the "forward" prop to pass an object of props to be attached to the component.
+
+<Link to="Route Name" forward={{ className: "test" }}>`);
+      }
     }
+
     const routeName = name || to;
+
     const Anchor: React.ReactType = anchor ? anchor : "a";
     const href: string = router.history.toHref({
       hash,
@@ -122,9 +142,14 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
       pathname: routeName ? router.route.pathname(routeName, params) : ""
     });
 
+    const additionalProps = {
+      ...rest,
+      ...forward
+    };
+
     return (
       <Anchor
-        {...rest}
+        {...additionalProps}
         onClick={this.clickHandler}
         href={href}
         ref={forwardedRef}